+++ src/main/java/pl/application/domain/guest/GuestService.java
@@ -36,11 +36,11 @@ public class GuestService {
         this.repository.readAll();
     }
 
-    public void removeGuest(int id) {
+    public void removeGuest(long id) {
         this.repository.remove(id);
     }
 
-    public void editGuest(int id, String firstName, String lastName, int age, boolean isMale) {
+    public void editGuest(long id, String firstName, String lastName, int age, boolean isMale) {
         Gender gender = Gender.FEMALE;
         if (isMale) {
             gender = Gender.MALE;
@@ -48,7 +48,7 @@ public class GuestService {
         this.repository.edit(id,firstName,lastName,age,gender);
     }
 
-    public Guest getGuestById(int id) {
+    public Guest getGuestById(long id) {
         return this.repository.findById(id);
     }
