+++ src/main/java/pl/application/domain/reservation/Reservation.java
@@ -7,13 +7,13 @@ import pl.application.domain.room.Room;
 import java.time.LocalDateTime;
 
 public class Reservation {
-    private final int id;
+    private final long id;
     private final Room room;
     private final Guest guest;
     private final LocalDateTime from;
     private final LocalDateTime to;
 
-    public Reservation(int id, Room room, Guest guest, LocalDateTime from, LocalDateTime to) {
+    public Reservation(long id, Room room, Guest guest, LocalDateTime from, LocalDateTime to) {
         this.id = id;
         this.room = room;
         this.guest = guest;
@@ -29,7 +29,7 @@ public class Reservation {
                 this.to.toString(),
                 System.getProperty("line.separator"));
     }
-    public int getId() {
+    public long getId() {
         return id;
     }
